Feature Checkout Credit Charge
This is the first stage in creating the payment functionality using
Stripe payments.  Only features are available for purchase and tests
for failed payments, successful payments and incorrect credit card
numbers are complete.

The second stage will be connecting the process
with Feature/Issue Voting so a user can only vote for a Feature
if they have made at least one payment for the Feature